Move main to bootstrap module. Update Dockerfile Update Controller tests to use WebMvcTest
