+++ food-quiz-api/src/test/java/com/guinardsolutions/foodquiz/api/TestConfig.java
@@ -1,0 +1,19 @@
+package com.guinardsolutions.foodquiz.api;
+
+import com.guinardsolutions.foodquiz.api.controller.QuizController;
+import org.springframework.boot.autoconfigure.EnableAutoConfiguration;
+import org.springframework.boot.data.jpa.autoconfigure.DataJpaRepositoriesAutoConfiguration;
+import org.springframework.boot.hibernate.autoconfigure.HibernateJpaAutoConfiguration;
+import org.springframework.boot.jdbc.autoconfigure.DataSourceAutoConfiguration;
+import org.springframework.context.annotation.ComponentScan;
+import org.springframework.context.annotation.Configuration;
+
+@Configuration
+@EnableAutoConfiguration(exclude = {
+        DataSourceAutoConfiguration.class,
+        HibernateJpaAutoConfiguration.class,
+        DataJpaRepositoriesAutoConfiguration.class
+})
+@ComponentScan(basePackageClasses = QuizController.class)
+public class TestConfig {
+}
