+++ food-quiz-api/src/test/java/com/guinardsolutions/foodquiz/api/controller/QuizControllerTests.java
@@ -1,13 +1,12 @@
 package com.guinardsolutions.foodquiz.api.controller;
 
+import com.guinardsolutions.foodquiz.api.TestConfig;
 import com.guinardsolutions.foodquiz.application.port.in.QuizResponse;
 import com.guinardsolutions.foodquiz.application.usecase.StartQuizUseCase;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.jdbc.test.autoconfigure.AutoConfigureTestDatabase;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.boot.webmvc.test.autoconfigure.AutoConfigureMockMvc;
-import org.springframework.test.context.ActiveProfiles;
+import org.springframework.boot.webmvc.test.autoconfigure.WebMvcTest;
+import org.springframework.test.context.ContextConfiguration;
 import org.springframework.test.context.bean.override.mockito.MockitoBean;
 import org.springframework.test.web.servlet.MockMvc;
 
@@ -18,11 +17,9 @@ import static org.springframework.test.web.servlet.request.MockMvcRequestBuilder
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.jsonPath;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
-@ActiveProfiles("test")
-@SpringBootTest
-@AutoConfigureMockMvc
-@AutoConfigureTestDatabase(replace = AutoConfigureTestDatabase.Replace.NONE)
-class QuizEntityControllerTests {
+@WebMvcTest(controllers = QuizController.class)
+@ContextConfiguration(classes = TestConfig.class)
+class QuizControllerTests {
 
     @Autowired
     private MockMvc mockMvc;
