+++ food-quiz-bootstrap/src/main/java/com/guinardsolutions/foodquiz/FoodQuizApplication.java
@@ -1,4 +1,4 @@
-package com.guinardsolutions.foodquiz.api;
+package com.guinardsolutions.foodquiz;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
