+++ food-quiz-bootstrap/src/main/resources/application.yaml
