+++ food-quiz-bootstrap/src/test/resources/application-test.yaml
